backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
vosk==0.3.45
moviepy==1.0.3
python-jose==3.3.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
requests==2.31.0
tqdm==4.66.1
slowapi==0.1.8
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
redis==5.0.1
celery==5.3.6
memory-profiler==0.61.0
aioredis==2.0.1
flower==2.0.1
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
starlette==0.27.0
ffmpeg-python==0.2.0
numpy==1.26.2
Pillow==10.1.0 